pyproject.toml
[build-system]
requires = ["setuptools>=71", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "myconex"
version = "0.1.0"
description = "Distributed AI mesh system inspired by biological fungal networks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai", "mesh", "distributed", "llm", "ollama", "nats", "mdns"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Networking",
]
dependencies = [
    "nats-py>=2.7.2",
    "zeroconf>=0.131.0",
    "redis>=5.0.4",
    "qdrant-client>=1.9.1",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "psutil>=5.9.8",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "click>=8.1.7",
    "rich>=13.7.1",
    "loguru>=0.7.2",
    "discord.py>=2.7.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]
gpu = [
    "gputil>=1.4.0",
]
ai = [
    "ollama>=0.2.1",
    "openai>=1.30.0",
    "sentence-transformers>=2.7.0",
]

[project.scripts]
myconex = "main:cli"

[project.urls]
Homepage = "https://github.com/myconex/myconex"
Repository = "https://github.com/myconex/myconex"
Issues = "https://github.com/myconex/myconex/issues"

# ─── Ruff (linter + formatter) ────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    "mobile",
    "docs",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E501",   # line too long — handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]   # re-exports are intentional

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# ─── MyPy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["mobile/", "docs/", "venv/"]

# ─── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["core", "orchestration"]
omit = ["tests/*", "mobile/*", "docs/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
